move of test for transition to use kafkaNodePool into FG
